Iterate on the requirements doc.
We definitely need these concise definitions of goals (and nongoals;
if anything, I think we might do well to have more of those!), but we
have indeed refined a few things in the last 8 months or so.

This is mostly addressing the huge batch of comments I left in the
vicinity of
https://github.com/ipld/specs/pull/146#pullrequestreview-268375431 .
Particularly tricky is that some of our most interesting and useful
ideas over the past few months *relax* some of the rules -- and so
we need to carefully come up with a way to describe that: it is
mostly recontextualizations consistent with the same underlying
philosophies... but boy is that hard to express in text sometimes!