Foundations: focus link descriptions more.
I've written a new introduction of linking for that heading which
focuses in on the parts we're sure about (and their upsides).

Some things seem more like wishlists or are unratified (and that in
some cases because the complexity of implications hasn't been
sufficiently explored), and I removed much of that content.
(Likely a good place to explore these things further would be by making
new files in the "exploration reports" design directories.)

Some of the wishlist content has also become fairly solved rather than
being todos since the earlier rounds of this document's life (namely,
some of the wishes for "slicing" can now actually be done using
Selectors!  which is fairly awesome), so that can also be dropped.

This addresses several earlier reviews:

- https://github.com/ipld/specs/pull/146#discussion_r308751905
- https://github.com/ipld/specs/pull/146#discussion_r308752488
- https://github.com/ipld/specs/pull/146#discussion_r308827938