Discussion: IPLD Data Model
I add a document to the exploration report repo, though it's not really
a full exploration report. It should be used as a discussion starting
point about the IPLD Data Model.

I decided to push it as a PR, so that we can also have inline comments.

This isn't really meant to be merged. Hopefully the outcome of this
discussion will lead to something that then can be merged.